refactor(interp): simplify interpreter call graphs
Separates interpreter logic into multiple recursive blocks, and nests
non-recursive functions. The point is to reduce the number of top-level
mutually recursive functions to simplify the call graph and support
better modularization in the future.
Long helper functions for `eval_exp` and `eval_prem`, `eval_instr` are
left as-is for readability, but is subject to change.