refactor(test): modularize test runner
The test runner is split up into a common library, and different
modules for the corresponding SpecTec modules that they test.

Each test can be run separately with `dune build @test/elab/runtest`.
Corresponding `make` commands are added for convenience, as well as
`test-quick` that runs only the fast tests and `test` that runs all
tests.